+++ common/src/main/java/com/kyanite/deeperdarker/registry/entities/custom/ai/nav/BetterWallClimberNavigation.java
@@ -47,14 +47,10 @@ public class BetterWallClimberNavigation extends GroundPathNavigation {
             super.tick();
         } else {
             if(this.pathToPosition != null) {
-                if(!this.pathToPosition.closerToCenterThan(this.mob.position(),
-                        Math.max(this.mob.getBbWidth(), 1.0D))
+                if(!this.pathToPosition.closerToCenterThan(this.mob.position(), Math.max(this.mob.getBbWidth(), 1.0D))
                         && (!(this.mob.getY() > (double) this.pathToPosition.getY())
-                        || !(new BlockPos(this.pathToPosition.getX(), this.mob.getY(),
-                        this.pathToPosition.getZ())).closerToCenterThan(this.mob.position(),
-                        Math.max(this.mob.getBbWidth(), 1.0D)))) {
-                    this.mob.getMoveControl().setWantedPosition(this.pathToPosition.getX(), this.pathToPosition.getY(),
-                            this.pathToPosition.getZ(), this.speedModifier);
+                        || !(new BlockPos(this.pathToPosition.getX(), (int) this.mob.getY(), this.pathToPosition.getZ())).closerToCenterThan(this.mob.position(), Math.max(this.mob.getBbWidth(), 1.0D)))) {
+                    this.mob.getMoveControl().setWantedPosition(this.pathToPosition.getX(), this.pathToPosition.getY(), this.pathToPosition.getZ(), this.speedModifier);
                 } else {
                     this.pathToPosition = null;
                 }
